// packages/graph/src/builder/types/state.d.ts
export type DirectAdapterLifetime<TAdapter> = TAdapter extends {
    lifetime: "singleton";
} ? "singleton" : TAdapter extends {
    lifetime: "scoped";
} ? "scoped" : TAdapter extends {
    lifetime: "transient";
} ? "transient" : InvalidLifetimeErrorMessage;